rename module files to avoid clashed with function names
