+++ src/plopp/plotting/inspector_plot.py
