+++ src/plopp/plotting/scatter3d_plot.py
