+++ src/plopp/plotting/scatter_plot.py
