+++ src/plopp/plotting/slicer_plot.py
