+++ src/plopp/plotting/super_plot.py
@@ -7,7 +7,7 @@ import scipp as sc
 
 from ..core.typing import FigureLike, Plottable
 from ..widgets import LineSaveTool
-from .slicer import SlicerPlot
+from .slicer_plot import SlicerPlot
 
 
 def superplot(
